src/test/java/onlineshop/controller/ProductControllerTest.java
package onlineshop.controller;

import com.fasterxml.jackson.databind.ObjectMapper;
import onlineshop.model.Product;
import onlineshop.service.ProductService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.math.BigDecimal;
import java.util.List;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.anyInt;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(ProductController.class)
class ProductControllerTest {

    @Autowired
    private MockMvc mockMvc;
    @Autowired
    private ObjectMapper objectMapper;

    @MockBean
    private ProductService productService;
    @MockBean
    private Product product;

    @BeforeEach
    public void setUp() {
        product = new Product(
                1,
                "Test product",
                "Test description",
                new BigDecimal( 99),
                10);
    }

    @Test
    public void testFindAll() throws Exception {
        String json = objectMapper.writeValueAsString(List.of(product));
        when(productService.findAll()).thenReturn(List.of(product));

        mockMvc.perform(get("/api/products"))
                .andExpect(status().isOk())
                .andExpect(content().string(json));
    }

    @Test
    public void testFindById() throws Exception {
        String json = objectMapper.writeValueAsString(product);
        when(productService.findById(anyInt())).thenReturn(product);

        mockMvc.perform(get("/api/products/{id}", 1))
                .andExpect(status().isOk())
                .andExpect(content().string(json));
    }

    @Test
    public void testCreateNewProduct() throws Exception {
        String json = objectMapper.writeValueAsString(product);
        when(productService.create(any(Product.class))).thenReturn(product);

        mockMvc.perform(post("/api/products")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(json))
                .andExpect(status().isOk())
                .andExpect(content().string(json));
    }
}